# openapi.yml
openapi: 3.0.3
info:
  title: Luga Store API
  version: 1.0.0
  description: API documentation for the Luga Store backend, listing all customer, shop manager/owner, and administrator endpoints.
security:
  - CookieAuth: []
paths:
...
  /profile/wishlist/{productId}:
    post:
      tags: [Wishlist]
      summary: Add a product to the wishlist (with unique composite check)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product added to wishlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    delete:
      tags: [Wishlist]
      summary: Remove a product from the wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Product removed from wishlist
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string

  # --- ORDERS (Checkout & History) ---
...
  /users/admins:
    get:
      tags: [Users (Admin)]
      summary: Get paginated list of administrators
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of admins retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'

    post:
      tags: [Users (Admin)]
      summary: Invite a new administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteAdminRequest'
      responses:
        '200':
          description: Invitation sent successfully
  /users/admins/invited:
    get:
      tags: [Users (Admin)]
      summary: Get paginated list of invited administrators
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of invited admins retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvitedUserList'
  /users/customers:
    get:
      tags: [Users (Admin)]
      summary: Get paginated list of customers
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of customers retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
  /users/{id}:
    get:
      tags: [Users (Admin)]
      summary: Retrieve a single user profile by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    patch:
      tags: [Users (Admin)]
      summary: Update a user's details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
      '200':
        description: User updated successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /users/admins/{id}:
    delete:
      tags: [Users (Admin)]
      summary: Delete an administrator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Admin deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /users/customers/{id}:
    delete:
      tags: [Users (Admin)]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Customer deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /users/admins/{id}/resend-invitation:
    post:
      tags: [Users (Admin)]
      summary: Resend admin invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Invitation resent successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /users/admins/{id}/invitation:
    delete:
      tags: [Users (Admin)]
      summary: Delete pending admin invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Invitation deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string

  # --- SHOPS (Admin/Owner/Manager) ---
  /shops:
    get:
      tags: [Shops]
      summary: Get paginated list of shops (Admin only)
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of shops

    post:
      tags: [Shops]
      summary: Create a new shop (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertShopRequest'
      responses:
        '200':
          description: Shop created successfully
  /shops/{id}:
    get:
      tags: [Shops]
      summary: Get details of a single shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop details

    put:
      tags: [Shops]
      summary: Update a shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertShopRequest'
      responses:
        '200':
          description: Shop updated successfully

    delete:
      tags: [Shops]
      summary: Delete a shop (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Shop deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{id}/status:
    patch:
      tags: [Shops]
      summary: Toggle active status of a shop (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopActiveStatusRequest'
      responses:
        '200':
          description: Shop active status updated successfully

  # --- SHOP OWNERS ---
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
...
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        avatarFileName:
          type: string
          nullable: true
        isEmailConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InvitedUserDto:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    PaginatedUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalCount:
          type: integer
        pageIndex:
          type: integer
        pageSize:
          type: integer
    PaginatedInvitedUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvitedUserDto'
        totalCount:
          type: integer
        pageIndex:
          type: integer
        pageSize:
          type: integer
...
    InviteAdminRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
    UpsertShopRequest:
      type: object
      required: [name, subdomainSlug]
      properties:
        name:
          type: string
        subdomainSlug:
          type: string
        description:
          type: string
        logoFileName:
          type: string
        bannerFileName:
          type: string
    UpdateShopActiveStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
...